\name{ExperimentHubMetadata-class}
\docType{class}

% Class:
\alias{class:ExperimentHubMetadata}
\alias{ExperimentHubMetadata-class}
\alias{ExperimentHubMetadata}

% Constructors:
\alias{ExperimentHubMetadata}

% Show:
\alias{show,ExperimentHubMetadata-method}


\title{Class \link{ExperimentHubMetadata} objects and methods}

\description{

  \link{ExperimentHubMetadata} is used to represent record(s) in the
  server data base.

}

\usage{
ExperimentHubMetadata(ExperimentHubRoot=NA_character_, 
        BiocVersion=biocVersion(),
        SourceUrl=NA_character_, 
        SourceType=NA_character_, 
        SourceVersion=NA_character_, 
        SourceLastModifiedDate=as.POSIXct(NA_character_), 
        SourceMd5=NA_character_, 
        SourceSize=NA_real_,
        DataProvider=NA_character_, 
        Title=NA_character_, 
        Description=NA_character_,
        Maintainer=NA_character_, 
        Species=NA_character_, 
        TaxonomyId=NA_integer_, 
        Genome=NA_character_, 
        Tags=NA_character_,
        RDataClass=NA_character_, 
        RDataDateAdded=as.POSIXct(NA_character_),
        RDataPath=NA_character_,
        Coordinate_1_based=TRUE,
        Notes=NA_character_,
        DispatchClass=NA_character_,
        PreparerClass=NA_character_,
        Location_Prefix='http://s3.amazonaws.com/experimenthub/')
}

\details{
  Instances of this class are generated by a call to \code{addResources}.
}

\arguments{
    \item{ExperimentHubRoot}{
      \code{character(1)} Prefix of the local path to resources to be added 
      to ExperimentHub. Currently not used.
    }
    \item{SourceUrl}{
      \code{character(1)} TBD 
    }
    \item{SourceType}{
      \code{character(1)} TBD 
    }
    \item{SourceVersion}{
      \code{character(1)} TBD 
    }
    \item{SourceLastModifiedDate}{
      \code{POSIXct()} Date when resource was last modified
    }
    \item{SourceMd5}{
      \code{character(1)} md5 hash of original file
    }
    \item{SourceSize}{
      \code{numeric(1)} Number of bytes in original file
    }
    \item{DataProvider}{
      \code{character(1)} Original data provider 
    }
    \item{Title}{
      \code{character(1)} Title for this resource
    }
    \item{Description}{
      \code{character(1)} Description of the resource
    }
    \item{Species}{
      \code{character(1)} Species name
    }
    \item{TaxonomyId}{
      \code{character(1)} NCBI code
    }
    \item{Genome}{
      \code{character(1)} Name of genome build
    }
    \item{Tags}{
      \code{character()} Free-form tags for search. biocViews from DESCRIPTION 
        are added to the \code{Tags}.
    }
    \item{RDataClass}{
      \code{character(1)} Class of derived object (e.g. \sQuote{GRanges})
    }
    \item{RDataDateAdded}{
      \code{POSIXct()} Date added to ExperimentHub; used to determine snapshots.
    }
    \item{RDataPath}{
      \code{character(1)} file path to serialized form
    }
    \item{Maintainer}{
      \code{character(1)} Maintainer name and email address, 
      \sQuote{A Maintainer \url{a.maintainer@email.addr}}
    }
    \item{BiocVersion}{
      \code{character()} Under which resource was built
    }
    \item{Coordinate_1_based}{
      \code{logical(1)} Do coordinates start with 1 or 0?
    }
    \item{Location_Prefix}{
      \code{character(1)} URL location of AWS S3 bucket.
    }
    \item{Notes}{
      \code{character()} Notes about the resource.
    }
    \item{DispatchClass}{
      \code{character(1)} Used internally.
    }
    \item{PreparerClass}{
      \code{character(1)} Used internally.
    }
}

\value{
  A \link{ExperimentHubMetadata} object.
}

\seealso{
  \itemize{
    \item \link{addResources}
    \item \link[ExperimentHub]{ExperimentHub}
  }
}

\examples{
## TODO
}

\keyword{classes}
